refactored configuration structure

+++ joule/models/stream.py
@@ -1,6 +1,7 @@
 from sqlalchemy.orm import relationship
 from sqlalchemy import (Column, Integer, String,
                         Boolean, Enum, ForeignKey)
+from sqlalchemy.dialects.postgresql import BIGINT
 from typing import List, Dict, TYPE_CHECKING
 import configparser
 import enum
@@ -56,7 +57,7 @@ class Stream(Base):
 
     KEEP_ALL = -1
     KEEP_NONE = 0
-    keep_us: int = Column(Integer, default=KEEP_ALL)
+    keep_us: int = Column(BIGINT, default=KEEP_ALL)
 
     description: str = Column(String)
     folder_id: int = Column(Integer, ForeignKey('metadata.folder.id'))
